+++ engineer/exceptions.py
@@ -6,6 +6,10 @@ class ThemeNotFoundException(Exception):
     pass
 
 
+class ThemeManifestError(Exception):
+    pass
+
+
 class ThemeDirectoryNotFoundException(Exception):
     def __init__(self, directory, *args, **kwargs):
         super(ThemeDirectoryNotFoundException, self).__init__(*args, **kwargs)