+++ engineer/themes.py
@@ -1,4 +1,5 @@
 # coding=utf-8
+from jinja2.loaders import FileSystemLoader
 import yaml
 from path import path
 from zope.cachedescriptors import method
@@ -30,12 +31,20 @@ class Theme(object):
         self.self_contained = kwargs.get('self_contained', True)
         self.static_root = path(kwargs.get('static_root', self.root_path / 'static/')).abspath()
         self.template_root = path(kwargs.get('template_root', self.root_path / 'templates')).abspath()
+        self.template_dirs = [self.template_root]
 
-        if 'templates' in kwargs:
-            self.templates = dict((k, self.theme_path(v)) for (k, v) in kwargs['templates'].iteritems())
-        else:
-            self.templates = dict((p.namebase, 'theme/%s' % p.name) for p in
-                self.template_root.walkfiles(pattern='*.html'))
+        # set up mappings for any additional content
+        self.content_mappings = {}
+        if 'copy_content' in kwargs:
+            for item, target in iter(kwargs['copy_content']):
+                item_path = path(item)
+                source = path(self.root_path / item_path).abspath()
+                if target is None:
+                    target = item_path
+                self.content_mappings[source] = target
+
+        if 'template_dirs' in kwargs:
+            self.template_dirs.extend([path(self.root_path / t).abspath() for t in kwargs['template_dirs']])
 
         # set the default theme settings values
         default_settings = kwargs.get('settings', None)
@@ -54,6 +63,10 @@ class Theme(object):
     @property
     def TEMPLATE_DIR(self):
         return self.template_root
+
+    @property
+    def template_loader(self):
+        return FileSystemLoader(self.template_dirs)
 
     def theme_path(self, template):
         if (self.template_root / template).abspath().exists():